docs: Update getting started guide with language greetings
* This is a twist to the hello world that makes it more interesting,
  and illustrates session memory in a good way
* Separate helloworld-agent sample instead of archetype
* The archetype will be removed in separate PR to make it easy to
  revert that if we change our mind
