+++ .github/workflows/ci.yml
@@ -377,6 +377,7 @@ jobs:
 
           - { sample: ask-akka-agent, it: false }
           - { sample: multi-agent, it: false, test: true }
+          - { sample: helloworld-agent, it: false, test: true }
 
     steps:
       - name: Checkout
