+++ docs/src/modules/getting-started/pages/samples.adoc
@@ -12,7 +12,7 @@ Illustrates how to create embeddings for vector databases, how to consume LLMs a
 +
 xref:getting-started:ask-akka-agent/index.adoc[Step-by-step guide]
 +
-link:https://github.com/akka-samples/ask-akka-agent[Github Repository],
+link:https://github.com/akka-samples/ask-akka-agent[Github Repository, window="new"],
 Level: Intermediate
 
 
@@ -20,42 +20,42 @@ AI agent that creates personalized travel itineraries::
 
 Illustrates reliable interaction with an LLM using a workflow. Entities are used for durable state of user preferences and generated trips.
 +
-link:https://github.com/akka-samples/travel-agent[Github Repository],
+link:https://github.com/akka-samples/travel-agent[Github Repository, window="new"],
 Level: Beginner
 
 AI agent that leverages an LLM to process medical discharge summaries::
 
 It assigns tags to the summaries, while also enabling human verification and comparative analysis. Interactions from a workflow with an agent using OpenAI LLM.
 +
-link:https://github.com/akka-samples/medical-tagging-agent[Github Repository], Level: Intermediate
+link:https://github.com/akka-samples/medical-tagging-agent[Github Repository, window="new"], Level: Intermediate
 
 
 AI agent that creates release notes summaries::
 
 Every time there is a release from set up GitHub repositories. Interactions with Anthropic Claude from an agent and using tools to retrieve detailed information from GitHub. Entities are used for storing release summaries. Timed action looks for new releases periodically and creates the summary using the LLM.
 +
-link:https://github.com/akka-samples/changelog-agent[Github Repository], Level: Intermediate
+link:https://github.com/akka-samples/changelog-agent[Github Repository, window="new"], Level: Intermediate
 
 Agentic workflow for customer service::
 
 The real-estate customer service agent is demonstrating how to combine Akka features with an LLM model. It illustrates an agentic workflow for customer service. It processes incoming real-estate inquiries, analyzes the content to extract details, provides follow-up when needed and saves the collected information for future reference.
 +
-link:https://github.com/akka-samples/real-estate-cs-agent[Github Repository], Level: Intermediate
+link:https://github.com/akka-samples/real-estate-cs-agent[Github Repository, window="new"], Level: Intermediate
 
 Trip booking agent using tools::
 This app represents an agency that searches for flights and accommodations. It's composed by an LLM (Anthropic) using Spring AI and tools to find flights, accommodations and sending mails.
 +
-link:https://github.com/akka-samples/trip-agent[Github Repository], Level: Intermediate
+link:https://github.com/akka-samples/trip-agent[Github Repository, window="new"], Level: Intermediate
 
 Multi-agent, weather-aware travel agent and activity planner::
 Demonstrating how to build a multi-agent system using Akka and an LLM model. A workflow manages the user query process, handling the sequential steps of agent selection, plan creation, execution, and summarization.
 +
-link:https://github.com/akka-samples/multi-agent[Github Repository], Level: Advanced
+link:https://github.com/akka-samples/multi-agent[Github Repository, window="new"], Level: Advanced
 
 Analyze sensor data by an agent::
 AI agent that uses an LLM to analyze data from fitness trackers, medical records and other sensors. Integration with Fitbit and MongoDB Atlas.
 +
-link:https://github.com/akka-samples/healthcare-agent[Github Repository], Level: Intermediate
+link:https://github.com/akka-samples/healthcare-agent[Github Repository, window="new"], Level: Intermediate
 
 Shopping cart microservice::
 
@@ -63,7 +63,7 @@ Shows a very simple microservice implementing a shopping cart with an event-sour
 +
 xref:getting-started:build-and-deploy-shopping-cart.adoc[Step-by-step guide]
 +
-link:https://github.com/akka-samples/shopping-cart-quickstart[Github Repository], Level: Beginner
+link:https://github.com/akka-samples/shopping-cart-quickstart[Github Repository, window="new"], Level: Beginner
 
 
 A customer registry microservice::
@@ -76,20 +76,20 @@ Involve multiple Entities in a transaction through a Workflow::
 
 A funds transfer workflow between two wallets.
 +
-link:https://github.com/akka-samples/transfer-workflow-compensation[Github Repository], Level: Intermediate
+link:https://github.com/akka-samples/transfer-workflow-compensation[Github Repository, window="new"], Level: Intermediate
 
 
 A choreography saga for user registration::
 A user registration service implemented as a Choreography Saga.
 +
-link:https://github.com/akka-samples/choreography-saga-quickstart[Github Repository], Level: Advanced
+link:https://github.com/akka-samples/choreography-saga-quickstart[Github Repository, window="new"], Level: Advanced
 
 
 Akka Chess::
 
 a complete, resilient, automatically scalable, event-sourced chess game
 +
-link:https://github.com/akka-samples/akka-chess[Github Repository], Level: Advanced
+link:https://github.com/akka-samples/akka-chess[Github Repository, window="new"], Level: Advanced
 
 
 [sidebar]
