+++ samples/helloworld-agent/src/main/java/com/example/api/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * This module is for the public API of the service, where the classes that define endpoints of the service should live.
+ */
+package com.example.api;
