+++ samples/helloworld-agent/src/main/java/com/example/application/package-info.java
@@ -1,0 +1,5 @@
+/**
+ * This module is for the application layer of the service. It is where the internal Akka components like entities and
+ * views should be defined.
+ */
+package com.example.application;
