+++ samples/helloworld-agent/src/main/java/com/example/domain/package-info.java
@@ -1,0 +1,5 @@
+/**
+ * This module is for the domain model of the service. It is should contain plain Java classes that does not
+ * depend on the Akka APIs.
+ */
+package com.example.domain;
