+++ samples/helloworld-agent/src/main/resources/include-dev-loggers.xml
@@ -1,0 +1,13 @@
+<?xml version="1.0" encoding="utf-8"?>
+<!-- 
+     This is a logback include file. 
+
+     Add here the loggers you would like to have for your application during development and testing.
+     To define loggers for production mode use 'include-loggers.xml'
+
+     The loggers you define inside the <included> tag will be added to the final log configuration 
+     when the application runs. 
+ -->
+<included>
+    <logger name="com.example" level="INFO"/>
+</included>
