+++ samples/helloworld-agent/src/test/java/com/example/IntegrationTest.java
@@ -1,0 +1,22 @@
+package com.example;
+
+import org.junit.jupiter.api.Test;
+import akka.javasdk.testkit.TestKitSupport;
+
+
+
+/**
+ * This is a skeleton for implementing integration tests for an Akka application built with the Akka SDK.
+ * <p>
+ * It interacts with the components of the application using a componentClient
+ * (already configured and provided automatically through injection).
+ */
+public class IntegrationTest extends TestKitSupport {
+
+
+  @Test
+  public void test() throws Exception {
+    // implement your integration tests here by calling your
+    // components by using the `componentClient`
+  }
+}
