+++ samples/multi-agent/src/main/resources/application.conf
@@ -1,5 +1,6 @@
 akka.javasdk {
   agent {
+    # Other AI models can be configured, see https://doc.akka.io/java/agents.html#model
     model-provider = openai
 
     openai {
